Moving utils to a new utils file
