--- pointer.go
@@ -1,6 +1,0 @@
-package patcher
-
-// ptr returns a pointer to the value passed in.
-func ptr[T any](v T) *T {
-	return &v
-}
